dbt_transform_astro/requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://www.astronomer.io/docs/astro/runtime-image-architecture#provider-packages
astronomer-cosmos==1.4.0
dbt-snowflake==1.8.4  # Updated from 1.9.2
snowflake-connector-python==3.13.2  # Updated from 3.5.0
apache-airflow-providers-snowflake==5.1.0  # Latest version for Airflow 2.8.1, check pypi.org for updates





